+++ MediStock/Presentation/Catalog/AllMedicinesView.swift
@@ -6,10 +6,10 @@
 //
 
 import SwiftUI
+import Toasty
 
 struct AllMedicinesView: View {
     @EnvironmentObject var viewModel: CatalogViewModel
-    @EnvironmentObject var authenticationViewModel: AuthenticationViewModel
     @Environment(\.diContainer) private var container
     @State private var filterText: String = ""
     @State private var sortOption: SortOption = .none
@@ -70,7 +70,6 @@ struct AllMedicinesView: View {
                 .sheet(isPresented: $isPresentingAddMedicine) {
                     AddMedicineView()
                         .environmentObject(viewModel)
-                        .environmentObject(authenticationViewModel)
                 }
             }
         }
@@ -81,7 +80,7 @@ struct AllMedicinesView_Previews: PreviewProvider {
     static var previews: some View {
         AllMedicinesView()
             .environmentObject(CatalogViewModel(medicineStore: FirestoreMedicineStore(),
-                                                historyStore: FirestoreHistoryStore()))
-            .environmentObject(AuthenticationViewModel(authenticationService: FirebaseAuthenticationService()))
+                                                historyStore: FirestoreHistoryStore(authenticationService: FirebaseAuthenticationService())))
+            .environmentObject(ToastyManager())
     }
 }